Implement C based api
