+++ src/c/bindings.jl
@@ -1,0 +1,1 @@
+include("kinesis.jl")
